Move sheetInfo methods to common map fragment class
