+++ main/src/main/java/cgeo/geocaching/WaypointPopupFragment.java
@@ -5,7 +5,6 @@ import cgeo.geocaching.apps.navi.NavigationAppFactory;
 import cgeo.geocaching.databinding.WaypointPopupBinding;
 import cgeo.geocaching.location.Geopoint;
 import cgeo.geocaching.location.Units;
-import cgeo.geocaching.maps.MapUtils;
 import cgeo.geocaching.models.Geocache;
 import cgeo.geocaching.models.Waypoint;
 import cgeo.geocaching.sensors.GeoData;
@@ -74,7 +73,7 @@ public class WaypointPopupFragment extends AbstractDialogFragmentWithProximityNo
 
         if (waypoint == null) {
             Log.e("WaypointPopupFragment.init: unable to get waypoint " + waypointId);
-            MapUtils.sheetRemoveFragment((AbstractNavigationBarMapActivity) requireActivity());
+            ((AbstractNavigationBarMapActivity) requireActivity()).sheetRemoveFragment();
             return;
         }
 
@@ -121,7 +120,7 @@ public class WaypointPopupFragment extends AbstractDialogFragmentWithProximityNo
 
             binding.moreDetails.setOnClickListener(arg0 -> {
                 CacheDetailActivity.startActivity(getActivity(), geocode);
-                MapUtils.sheetRemoveFragment((AbstractNavigationBarMapActivity) requireActivity());
+                ((AbstractNavigationBarMapActivity) requireActivity()).sheetRemoveFragment();
             });
 
             final View view = getView();
