+++ main/src/main/java/cgeo/geocaching/maps/MapUtils.java
@@ -1,13 +1,8 @@
 package cgeo.geocaching.maps;
 
-import cgeo.geocaching.CacheDetailActivity;
-import cgeo.geocaching.CachePopupFragment;
 import cgeo.geocaching.CgeoApplication;
 import cgeo.geocaching.EditWaypointActivity;
 import cgeo.geocaching.R;
-import cgeo.geocaching.SwipeToOpenFragment;
-import cgeo.geocaching.WaypointPopupFragment;
-import cgeo.geocaching.activity.AbstractNavigationBarMapActivity;
 import cgeo.geocaching.activity.ActivityMixin;
 import cgeo.geocaching.apps.navi.NavigationAppFactory;
 import cgeo.geocaching.connector.internal.InternalConnector;
@@ -33,7 +28,6 @@ import cgeo.geocaching.storage.extension.OneTimeDialogs;
 import cgeo.geocaching.ui.CoordinatesFormatSwitcher;
 import cgeo.geocaching.ui.dialog.Dialogs;
 import cgeo.geocaching.ui.dialog.SimplePopupMenu;
-import cgeo.geocaching.unifiedmap.UnifiedMapViewModel;
 import cgeo.geocaching.utils.AndroidRxUtils;
 import cgeo.geocaching.utils.ClipboardUtils;
 import cgeo.geocaching.utils.FilterUtils;
@@ -46,28 +40,19 @@ import android.app.Activity;
 import android.content.ActivityNotFoundException;
 import android.content.ComponentName;
 import android.content.Intent;
-import android.content.res.Resources;
 import android.graphics.Point;
 import android.text.Html;
 import android.text.Spanned;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.FrameLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.DrawableRes;
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.WorkerThread;
 import androidx.appcompat.app.AlertDialog;
-import androidx.coordinatorlayout.widget.CoordinatorLayout;
 import androidx.core.content.res.ResourcesCompat;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -78,15 +63,9 @@ import java.util.Set;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 
-import com.google.android.material.bottomsheet.BottomSheetBehavior;
-import com.google.android.material.sidesheet.SideSheetBehavior;
-import com.google.android.material.sidesheet.SideSheetCallback;
 import org.apache.commons.lang3.StringUtils;
 
 public class MapUtils {
-
-    private static final String TAG_MAPDETAILS_FRAGMENT = "mapdetails_fragment";
-    private static final String TAG_SWIPE_FRAGMENT = "swipetoopen_fragment";
 
     private MapUtils() {
         // should not be instantiated
@@ -361,134 +340,4 @@ public class MapUtils {
         menu.addItemClickListener(uniqueId, clickAction::call);
     }
 
-    // ---------------------------------------------------------------------------------------------
-    // Handling of cache/waypoint details fragments
-
-    public static void sheetShowDetails(final AbstractNavigationBarMapActivity activity, @Nullable final UnifiedMapViewModel.SheetInfo sheetInfo) {
-        if  (sheetInfo == null || StringUtils.isBlank(sheetInfo.geocode)) {
-            return;
-        }
-        if (sheetInfo.waypointId <= 0) {
-            sheetConfigureFragment(CachePopupFragment.newInstance(sheetInfo.geocode), activity, () -> CacheDetailActivity.startActivity(activity, sheetInfo.geocode));
-        } else {
-            sheetConfigureFragment(WaypointPopupFragment.newInstance(sheetInfo.geocode, sheetInfo.waypointId), activity, () -> CacheDetailActivity.startActivity(activity, sheetInfo.geocode));
-        }
-    }
-
-    private static void sheetConfigureFragment(final Fragment fragment, final AbstractNavigationBarMapActivity activity, final Runnable onUpSwipeAction) {
-        final FrameLayout fl = activity.findViewById(R.id.detailsfragment);
-        final ViewGroup.LayoutParams params = fl.getLayoutParams();
-        final CoordinatorLayout.Behavior<?> behavior = ((CoordinatorLayout.LayoutParams) params).getBehavior();
-        final boolean isBottomSheet = behavior instanceof BottomSheetBehavior;
-
-        final FragmentTransaction ft = activity.getSupportFragmentManager().beginTransaction();
-        final SwipeToOpenFragment swipeToOpenFragment = isBottomSheet ? new SwipeToOpenFragment() : null;
-        if (isBottomSheet) {
-            ft.replace(R.id.detailsfragment, swipeToOpenFragment, TAG_SWIPE_FRAGMENT);
-            ft.add(R.id.detailsfragment, fragment, TAG_MAPDETAILS_FRAGMENT);
-        } else {
-            ft.replace(R.id.detailsfragment, fragment, TAG_MAPDETAILS_FRAGMENT);
-        }
-        ft.commit();
-
-        if (isBottomSheet) { // portrait mode uses BottomSheet
-            final BottomSheetBehavior<FrameLayout> b = BottomSheetBehavior.from(fl);
-            b.setHideable(true);
-            b.setSkipCollapsed(false);
-            b.setPeekHeight(0); // temporary set to 0 to avoid bumping. Gets updated once view is loaded.
-            b.setState(BottomSheetBehavior.STATE_COLLAPSED);
-
-            ft.runOnCommit(() -> {
-                final View view = fragment.requireView();
-                // make bottom sheet fill whole screen
-                swipeToOpenFragment.requireView().setMinimumHeight(Resources.getSystem().getDisplayMetrics().heightPixels);
-                // set the height of collapsed state to height of the details fragment
-                view.getViewTreeObserver().addOnGlobalLayoutListener(() -> b.setPeekHeight(view.getHeight()));
-            });
-
-            final BottomSheetBehavior.BottomSheetCallback callback = new BottomSheetBehavior.BottomSheetCallback() {
-                @Override
-                public void onStateChanged(@NonNull final View bottomSheet, final int newState) {
-                    if (newState == BottomSheetBehavior.STATE_HIDDEN) {
-                        sheetRemoveFragment(activity);
-                    }
-                    if (newState == BottomSheetBehavior.STATE_EXPANDED && onUpSwipeAction != null) {
-                        onUpSwipeAction.run();
-                        ActivityMixin.overrideTransitionToFade(activity);
-                    }
-                }
-
-                @Override
-                public void onSlide(@NonNull final View bottomSheet, final float slideOffset) {
-                    swipeToOpenFragment.setExpansion(slideOffset, fragment.getView());
-                }
-            };
-
-            b.addBottomSheetCallback(callback);
-            swipeToOpenFragment.setOnStopCallback(() -> b.removeBottomSheetCallback(callback));
-        } else { // landscape mode uses SideSheet
-            final SideSheetBehavior<FrameLayout> b = SideSheetBehavior.from(fl);
-            b.setState(SideSheetBehavior.STATE_EXPANDED);
-            final SideSheetCallback[] callbackStore = new SideSheetCallback[] { null };
-
-            final SideSheetCallback callback = new SideSheetCallback() {
-                @Override
-                public void onStateChanged(@NonNull final View sheet, final int newState) {
-                    if (newState == SideSheetBehavior.STATE_HIDDEN) {
-                        sheetRemoveFragment(activity);
-                        if (callbackStore[0] != null) {
-                            b.removeCallback(callbackStore[0]);
-                        }
-                    }
-                }
-
-                @Override
-                public void onSlide(@NonNull final View sheet, final float slideOffset) {
-                    // nothing
-                }
-            };
-            callbackStore[0] = callback;
-            b.addCallback(callback);
-        }
-
-        fl.setVisibility(View.VISIBLE);
-    }
-
-    /** removes fragment and view for mapdetails view; returns true, if view got removed */
-    public static boolean sheetRemoveFragment(@NonNull final AbstractNavigationBarMapActivity activity) {
-        activity.clearSheetInfo();
-        final FragmentManager fm = activity.getSupportFragmentManager();
-        final Fragment f1 = fm.findFragmentByTag(TAG_MAPDETAILS_FRAGMENT);
-        if (f1 != null) {
-            fm.beginTransaction().remove(f1).commit();
-        }
-        final Fragment f2 = fm.findFragmentByTag(TAG_SWIPE_FRAGMENT);
-        if (f2 != null) {
-            fm.beginTransaction().remove(f2).commit();
-        }
-        final FrameLayout v = activity.findViewById(R.id.detailsfragment);
-        if (v != null && v.getVisibility() != View.GONE) {
-
-            final CoordinatorLayout.Behavior<?> behavior = ((CoordinatorLayout.LayoutParams) v.getLayoutParams()).getBehavior();
-            if (behavior instanceof BottomSheetBehavior) {
-                final BottomSheetBehavior<FrameLayout> b = BottomSheetBehavior.from(v);
-                b.setState(BottomSheetBehavior.STATE_HIDDEN); // close correctly as it will otherwise conflict with up-swipe behaviour implementation
-            }
-
-            v.setVisibility(View.GONE);
-            return true;
-        }
-        return false;
-    }
-
-    // manage onStart lifecycle for SheetInfo:
-    // - read current value
-    // - reset details fragment
-    // - restore current value
-    // - open new sheet (if non-empty)
-    public static void sheetManageLifecycleOnStart(final AbstractNavigationBarMapActivity activity, @Nullable final UnifiedMapViewModel.SheetInfo sheetInfo, @NonNull final Runnable clearSheetInfo, @NonNull final Action1<UnifiedMapViewModel.SheetInfo> setSheetInfo) {
-        MapUtils.sheetRemoveFragment(activity);
-        setSheetInfo.call(sheetInfo);
-        MapUtils.sheetShowDetails(activity, sheetInfo);
-    }
 }
