+++ main/src/main/java/cgeo/geocaching/maps/google/v2/GoogleMapView.java
@@ -8,7 +8,6 @@ import cgeo.geocaching.location.Viewport;
 import cgeo.geocaching.maps.CGeoMap;
 import cgeo.geocaching.maps.DistanceDrawer;
 import cgeo.geocaching.maps.MapProviderFactory;
-import cgeo.geocaching.maps.MapUtils;
 import cgeo.geocaching.maps.ScaleDrawer;
 import cgeo.geocaching.maps.interfaces.GeneralOverlay;
 import cgeo.geocaching.maps.interfaces.GeoPointImpl;
@@ -139,7 +138,7 @@ public class GoogleMapView extends MapView implements MapViewImpl<GoogleCacheOve
         googleMap.setOnCameraIdleListener(this::recognizePositionChange);
         googleMap.setOnMapClickListener(latLng -> {
             if (activityRef.get() != null) {
-                if (MapUtils.sheetRemoveFragment(activityRef.get())) {
+                if (activityRef.get().sheetRemoveFragment()) {
                     return;
                 }
                 adaptLayoutForActionbar(HideActionBarUtils.toggleActionBar(activityRef.get()));
