+++ main/src/main/java/cgeo/geocaching/unifiedmap/UnifiedMapActivity.java
@@ -861,7 +861,7 @@ public class UnifiedMapActivity extends AbstractNavigationBarMapActivity impleme
                         .setOnDismissListener(d -> viewModel.longTapCoords.setValue(null))
                         .show();
             } else {
-                if (MapUtils.sheetRemoveFragment(this)) {
+                if (sheetRemoveFragment()) {
                     return;
                 }
                 mapFragment.adaptLayoutForActionbar(HideActionBarUtils.toggleActionBar(this));
@@ -927,10 +927,10 @@ public class UnifiedMapActivity extends AbstractNavigationBarMapActivity impleme
             // open popup for element
             if (routeItem.getType() == RouteItem.RouteItemType.GEOCACHE) {
                 viewModel.sheetInfo.setValue(new UnifiedMapViewModel.SheetInfo(routeItem.getGeocode(), 0));
-                MapUtils.sheetShowDetails(this, viewModel.sheetInfo.getValue());
+                sheetShowDetails(viewModel.sheetInfo.getValue());
             } else if (routeItem.getType() == RouteItem.RouteItemType.WAYPOINT && routeItem.getWaypointId() != 0) {
                 viewModel.sheetInfo.setValue(new UnifiedMapViewModel.SheetInfo(routeItem.getGeocode(), routeItem.getWaypointId()));
-                MapUtils.sheetShowDetails(this, viewModel.sheetInfo.getValue());
+                sheetShowDetails(viewModel.sheetInfo.getValue());
             }
         } else if (item.getRoute() != null) {
             // elevation charts for individual route and/or routes/tracks
@@ -963,7 +963,7 @@ public class UnifiedMapActivity extends AbstractNavigationBarMapActivity impleme
     }
 
     @Override
-    public void clearSheetInfo() {
+    protected void clearSheetInfo() {
         Log.e("clearSheetInfo");
         viewModel.sheetInfo.setValue(null);
     }
@@ -1019,7 +1019,7 @@ public class UnifiedMapActivity extends AbstractNavigationBarMapActivity impleme
         super.onStart();
 
         // close outdated details popups & restart sheet (if required)
-        MapUtils.sheetManageLifecycleOnStart(this, viewModel.sheetInfo.getValue(), this::clearSheetInfo, sheetInfo -> viewModel.sheetInfo.setValue(sheetInfo));
+        sheetManageLifecycleOnStart(viewModel.sheetInfo.getValue(), sheetInfo -> viewModel.sheetInfo.setValue(sheetInfo));
 
         // resume location access
         resumeDisposables.add(geoDirUpdate.start(GeoDirHandler.UPDATE_GEODIR));
